Monolithic/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
tensorflow>=2.5.0
kafka-python>=2.0.2
influxdb-client>=1.23.0
paho-mqtt>=1.6.1
scikit-learn>=0.24.0
